Modify CharacterDisplay and CPUMoveSelect
Changed CharacterDisplay method,
GetSystemHealthPercentage, to return a float instead.
Due to percentage rounding errors, fractionnal values
less than 1% of the maxhp are rounded up or down.
This causes an issue if hp is less than 1%, where that
systems hp is considered at 0%. Thus it would be
skipped over during TriageTargets.

TriageTargets altered to not use Array.Sort because
the wrong index values were being recorded. Added
comparisons so that it records new index when
the lowest healthPercentage is surpassed.

+++ Assets/Scripts/Core/CharacterDisplay.cs
@@ -151,11 +151,11 @@ public class CharacterDisplay : MonoBehaviour
         EnergyAffected?.Invoke(EnergyBarData);
     }
 
-    public int GetSystemHealthPercentage(int systemIndex)
+    public float GetSystemHealthPercentage(int systemIndex)
     {
         // 0 is main health; 1 is head; 2 is chest; 3 is LArm; 4 is RArm; 5 is Legs 
         int targetSystemMaxHealthIndex = (systemIndex * 2);
-        int healthPercent = (100 * HealthSystemData[targetSystemMaxHealthIndex + 1]) / HealthSystemData[targetSystemMaxHealthIndex];
+        float healthPercent = (100f * (float)HealthSystemData[targetSystemMaxHealthIndex + 1]) / (float)HealthSystemData[targetSystemMaxHealthIndex];
         return healthPercent;
     }
